+++ packages/vfm/types/remark.d.ts
@@ -1,20 +1,11 @@
-declare module '*.svg' {
-  const content: string;
-  export default content;
-}
-
 declare namespace NodeJS {
   interface ProcessEnv {
     [key: string]: string;
   }
 }
 
-declare module 'remark-rehype' {
-  import {Plugin} from 'unified';
-  const plugin: Plugin;
-  export default plugin;
-}
-
+declare module 'remark-math';
+declare module 'remark-breaks';
 declare module 'rehype-raw' {
   import {Plugin} from 'unified';
   const plugin: Plugin;
@@ -26,12 +17,27 @@ declare module 'rehype-stringify' {
   const plugin: Plugin;
   export default plugin;
 }
-
-declare module 'remark-math';
 declare module 'rehype-mathjax';
-declare module 'hast-util-find-and-replace';
 declare module 'hastscript';
-declare module 'hast-util-is-element';
-declare module 'remark-breaks';
-
 declare module 'mdast-util-to-hast/lib/all';
+declare module 'hast-util-find-and-replace';
+declare module 'hast-util-is-element';
+
+interface Tokenizer {
+  (
+    this: TokenizerInstance,
+    eat: Eat & {now: () => Point},
+    value: string,
+    silent?: boolean,
+  ): boolean | Node | void;
+  locator?: Locator;
+  onlyAtStart?: boolean;
+  notInBlock?: boolean;
+  notInList?: boolean;
+  notInLink?: boolean;
+}
+
+type TokenizerInstance = {
+  tokenizeBlock: (value: string) => Node | void;
+  tokenizeInline: (value: string, location: Point) => Node | void;
+};
